fix: abstract out actor state type convertion logic in state_migration
