+++ vm/state_migration/src/type_migrations/miner/mod.rs
@@ -1,0 +1,4 @@
+// Copyright 2019-2023 ChainSafe Systems
+// SPDX-License-Identifier: Apache-2.0, MIT
+
+mod state_v10_to_v11;
