+++ vm/state_migration/src/type_migrations/mod.rs
@@ -1,0 +1,5 @@
+// Copyright 2019-2023 ChainSafe Systems
+// SPDX-License-Identifier: Apache-2.0, MIT
+
+mod init;
+mod miner;
